Be more careful with --ignore-arch-all
Historically, dpkg-source would output "Architecture: any" on any source
package that built binaries for more than one architecture.  As of
dpkg-dev version 1.15.1, it spells out each architecture, so we end up
with architecture lines like

  Architecture: all amd64 i386 ia64

for a package that builds arch-all binaries, plus architecture-specific
binaries on amd64, i386, and ia64, but not on any other architectures.

This patch changes the ignore_arch_all logic in parse_sources() to
ignore a package with "all" in the architecture list only if it also
doesn't find the current architecture (either detected in the given
Packages file or specified by --architecture) in the architecture list.

Closes: #530595

Signed-off-by: John Wright <john.wright@hp.com>
Signed-off-by: Philipp Kern <pkern@debian.org>