src/utils.c: process line continuations in parse_package
Line continuations (space as first character in a line) were just skipped
with the previous implementation.  Now they are processed, except for
descriptions, which are still skipped.  This enables quinn-diff to read
overlong binary fields in Sources files.  It might also be desirable to
skip Files and Checksum-* fields, which usually use continuations too
but are not necessary for quinn-diff's operation.

Signed-off-by: Philipp Kern <pkern@debian.org>