src/parse_sources.c: rework source architecture handling
The old code used Packages-arch-specific as the authoriative data source
about which package should be built and where.  Nowadays we get more
information about the source architectures from the Sources file.  This
patch ensures that the information found there is taken as authoriative
unless overridden through Packages-arch-specific.  This gives the
maintainer more power and requires less changes to P-a-s with the drawback
that such architecture lists cannot be changed without an upload.

However this is not a problem we can solve nowadays because dpkg-dev bails
out when the current architecture is not found in the architecture list.
Due to the current buildd architecture there are not many people who would
then bring that build to completion.

Closes: #275835

Signed-off-by: Philipp Kern <pkern@debian.org>