src/parse_sources.c: handle arch:all more sanely
An arch:all package is buildable on $arch if the source package's
architecture list either contains $arch, all or "all $arch", i.e.
it's not if neither the current arch nor any are included and if
the arch list does not equal a single all.

Furthermore this patch strips the whitespace from the architecture
list, so the offsets in the debug output are no longer necessary.

Signed-off-by: Philipp Kern <pkern@debian.org>